sameold: reorg message
Move `sameold::message::event` → `sameold::message::phenomenon`

Move SignificanceLevel and errors to
`sameold::message::significance`

These moves do not impact the external API.
